+++ examples/ns/query.py
@@ -1,10 +1,12 @@
+"""the main API"""
 import typing as t
 import xml.etree.ElementTree
 from datetime import datetime
-from operator import attrgetter
+from functools import singledispatch
+from operator import attrgetter, methodcaller
 
-from gentools import map_return, map_send, map_yield, oneyield, reusable
-from toolz import compose, valfilter
+from gentools import (compose, map_return, map_send, map_yield, oneyield,
+                      reusable)
 
 import snug
 
@@ -12,7 +14,24 @@ from .load import registry as loads
 from .types import Departure, Journey, Station
 
 API_PREFIX = 'https://webservices.ns.nl/ns-api-'
-parse_request = compose(xml.etree.ElementTree.fromstring, attrgetter('data'))
+parse_request = compose(xml.etree.ElementTree.fromstring,
+                        attrgetter('content'))
+
+
+@singledispatch
+def dump_param(val):
+    """dump a query param value"""
+    return str(val)
+
+
+dump_param.register(datetime, methodcaller('strftime', '%Y-%m-%dT%H:%M'))
+
+
+def prepare_params(req: snug.Request) -> snug.Request:
+    """prepare request parameters"""
+    return req.replace(
+        params={key: dump_param(val) for key, val in req.params.items()
+                if val is not None})
 
 
 def basic_query(returns):
@@ -20,20 +39,20 @@ def basic_query(returns):
     return compose(
         reusable,
         map_send(parse_request),
-        map_yield(snug.prefix_adder(API_PREFIX)),
+        map_yield(prepare_params, snug.prefix_adder(API_PREFIX)),
         map_return(loads(returns)),
         oneyield,
     )
 
 
 @basic_query(t.List[Station])
-def stations():
+def stations() -> snug.Query[t.List[Station]]:
     """a list of all stations"""
     return snug.GET('stations-v2')
 
 
 @basic_query(t.List[Departure])
-def departures(station: str):
+def departures(station: str) -> snug.Query[t.List[Departure]]:
     """departures for a station"""
     return snug.GET('avt', params={'station': station})
 
@@ -46,16 +65,16 @@ def journey_options(origin:      str,
                     after:       t.Optional[int]=None,
                     time:        t.Optional[datetime]=None,
                     hsl:         t.Optional[bool]=None,
-                    year_card:   t.Optional[bool]=None):
+                    year_card:   t.Optional[bool]=None) -> (
+                        snug.Query[t.List[Journey]]):
     """journey recommendations from an origin to a destination station"""
-    return snug.GET('treinplanner',
-                    params=valfilter(lambda x: x is not None, {
-                        'fromStation':     origin,
-                        'toStation':       destination,
-                        'viaStation':      via,
-                        'previousAdvices': before,
-                        'nextAdvices':     after,
-                        'dateTime':        time,
-                        'hslAllowed':      hsl,
-                        'yearCard':        year_card,
-                    }))
+    return snug.GET('treinplanner', params={
+        'fromStation':     origin,
+        'toStation':       destination,
+        'viaStation':      via,
+        'previousAdvices': before,
+        'nextAdvices':     after,
+        'dateTime':        time,
+        'hslAllowed':      hsl,
+        'yearCard':        year_card,
+    })