+++ snug/__about__.py
@@ -1,4 +1,4 @@
-__version__ = '0.4.0'
+__version__ = '0.5.0'
 __author__ = 'Arie Bovenberg'
 __copyright__ = '2018, Arie Bovenberg'
 __description__ = 'A compact toolkit for wrapping web APIs'