+++ HISTORY.rst
@@ -4,6 +4,13 @@ Release history
 development
 +++++++++++
 
+0.5.0 (2018-01-30)
+++++++++++++++++++
+
+- improvements to docs
+- rename Request/Response data->content
+- ``Relation`` query class
+
 0.4.0 (2018-01-24)
 ++++++++++++++++++